apps/backend/requirements.txt
torch
transformers
sentencepiece
fastapi==0.123.3
uvicorn[standard]==0.38.0
sqlalchemy[asyncio]==2.0.44
asyncpg==0.31.0
pydantic-settings==2.12.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
httpx==0.27.0
# openai==1.54.0  <-- Removed
# anthropic==0.34.0 <-- Removed
